Bug fixes and removal of unnecessary code
Fixed a bug where the ArduinoFile doesn't wait for the host to raise /CMD before checking for an 0x55 on the bus, causing disk errors on some Lisas (but not mine for some reason). Also removed code related to the key-value store for Tom Stepleton's Selector program; after partially implementing it a few months ago, I realized that it would be way too slow on an Arduino and so I just removed it entirely.